replace stoll with stoi to ensure proper parsing

+++ src/main.cpp
@@ -2,12 +2,12 @@
 
 #include "../include/Game.h"
 
-int main(int argc, char *argv[]) {
+int main(int argc, char* argv[]) {
 
     try {
-        const int xSize = std::stoll(argv[1]);
-        const int ySize = std::stoll(argv[2]);
-        const int tileSize = std::stoll(argv[3]);
+        const int xSize = std::stoi(argv[1]);
+        const int ySize = std::stoi(argv[2]);
+        const int tileSize = std::stoi(argv[3]);
 
         const bool showWindow = argv[4];
